Use CLOCK_MONOTONIC to determine the clock_tick variable

+++ include/task.hpp
@@ -59,7 +59,7 @@ private:
     /* How many task are in this wheel. */
     int count;
 
-    unsigned int wheel_base;
+    unsigned long wheel_base;
 
     /* The actual wheels. This is a variation on having one wheel for
      * seconds, another for minutes and a third for hours - in this
